+++ python/GafferSceneUI/__init__.py
@@ -211,6 +211,7 @@ from . import DisplayUI
 from . import CatalogueUI
 from . import CatalogueSelectUI
 from . import ImageSelectionToolUI
+from . import CameraQueryUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.